Last updates to server side of task, added a lot of functionality includes many exception handlers, work with admin requests and a lot mini changes

+++ src/main/java/org/onishkoff/itmo/IS1/dto/model/response/CaveDto.java
@@ -7,7 +7,7 @@ import lombok.Data;
 @Builder
 public class CaveDto{
     private Long id;
-    private Long depth;
-    private Long numberOfTreasures;
+    private Double depth;
+    private Double numberOfTreasures;
 
 }
